GP-3194 MATLAB client build script now points to the current version of the previously missing jar.  Updated the pointer to the commons-logging jar as well, to the newer version of that jar.
Note:  I discovered that building the MATLAB client correctly involves first building the gpwebclient.jar through the build script (ant analysis), as the jar produced by that script will be included if it exists in the zip produced for the MATLAB client.  If the jar doesn't exist for inclusion, the MATLAB client zip will still be created, but will not function correctly.

git-svn-id: https://svn.broadinstitute.org/genepattern/trunk@9325 48a7a3f7-ad46-0410-b969-29950918272e