Server-side code for saving and loading a license with the PipelineDesigner
Note:  This code works, but assumes you know what you're doing.  If, for example, you save a license file that doesn't match our filename convention, then on load it shows that you have no license, since none was found onLoad matching the pattern.  That's because the code just calls the EULAManager to get the license.  Improvements still need to be made to get this working with our new way of declaring the license in the manifest.

git-svn-id: https://svn.broadinstitute.org/genepattern/trunk@11022 48a7a3f7-ad46-0410-b969-29950918272e